+++ lib/app/app.dart
@@ -8,3 +8,10 @@ export 'routes/routes.dart';
 final rootRouter = RootAppRouterDelegate();
 
 AppRouterDelegate get router => rootRouter.appState.activeRouter;
+
+mixin RouteInfo {
+  String get route;
+  void routeTo() {
+    router.push(url: route);
+  }
+}