+++ lib/app/modules/trait/tabs/buff_tab.dart
@@ -121,6 +121,7 @@ class _TraitBuffTabState extends State<TraitBuffTab> {
             _traits('UpBuffRate', buff.script!.UpBuffRateBuffIndiv!),
         ],
       ),
+      onTap: () => buff.routeTo(),
     );
   }
 }